fix: snapshot - pol mode indexing
git-svn-id: https://www.brockmann-consult.de/svn/os/smos-box/trunk@4206 d90d2d48-2b33-0410-a973-e72f85796a02

+++ smos-reader/src/main/java/org/esa/beam/dataio/smos/L1cBrowseSmosFile.java
@@ -53,19 +53,19 @@ public class L1cBrowseSmosFile extends L1cSmosFile {
 
     @Override
     public short getSnapshotBtData(int gridPointIndex, int fieldIndex, int polMode,
-                                   int snapshotId, short noDataValue) throws IOException {
+                                   long snapshotId, short noDataValue) throws IOException {
         return getBtData(gridPointIndex, polMode).getShort(fieldIndex);
     }
 
     @Override
     public int getSnapshotBtData(int gridPointIndex, int fieldIndex, int polMode,
-                                 int snapshotId, int noDataValue) throws IOException {
+                                 long snapshotId, int noDataValue) throws IOException {
         return getBtData(gridPointIndex, polMode).getInt(fieldIndex);
     }
 
     @Override
     public float getSnapshotBtData(int gridPointIndex, int fieldIndex, int polMode,
-                                   int snapshotId, float noDataValue) throws IOException {
+                                   long snapshotId, float noDataValue) throws IOException {
         return getBtData(gridPointIndex, polMode).getFloat(fieldIndex);
     }
